Improve Kubernetes health check script
Instead of testing the connection once, after waiting for 30 seconds,
the script now tests the connection every second for 30 seconds.
This ensures that the script will detect downtime during the deployment, not just after.
If deployment starts taking longer than 30 seconds, the value should be adjusted.

Changes to be committed:
	modified:   scripts/kubernetes-test.sh