Update GitHub CI workflow to ignore deployment and health check on non-main branches
To avoid deploying features that are still in development,
the GitHub CI workflow file has been updated to only run the
deployment job on the main branch.
Since the health check job is supposed to detect faulty
deployments, this job doesn't need to run if the deployment job
is disabled. Thus, this job has been disabled as well.

Changes to be committed:
	modified:   .github/workflows/workflow.yaml